Moved CompImageSection to separate file
